feat: Add Zig-Zag number generator
This commit adds a new file, Zig-zag.cpp, which contains the implementation of a function to generate Zig-Zag numbers up to a given limit. Zig-Zag numbers are generated using a breadth-first search (BFS) algorithm. The function takes an input number X and finds the largest Zig-Zag number that is less than or equal to X. The algorithm uses a queue to perform the BFS and checks if each number is a valid Zig-Zag number. The largest Zig-Zag number found is then returned.

The main function in the file takes an input number X from the user and calls the Zig-Zag number generator function to find the largest Zig-Zag number less than or equal to X. If X is 0, the largest Zig-Zag number is 0. The result is then printed to the console.

This feature is implemented to provide a utility for generating Zig-Zag numbers, which can be useful in certain mathematical and programming scenarios.